Support Undercurl and Overline, rename ReverseVideo → Reverse CrossedOut → StrikeOut
Remove BlinkSlow and BlinkRapid in favour of this. BlinkRapid is not
widely supported, so that's useless in the first place. And BlinkSlow is
somewhat widely supported, but a few disable it by default (kitty,
alacritty). It's also damn annoying, and you should probably never use
it.

The format for the extended Smulx is:

	1=regular, 2=double, 3=undercurl, 4=dotted, 5=dashed

But only undercurl is supported for now. We could, in theory, repurpose
the error bit to have 4 options for this. Maybe something for later; the
other underline styles don't seem too useful to me.

Rename ReverseVideo to be shorter, and CrossedOut to StrikeOut to be
more standard.