Fix GetVersion and PrintVersion with "go install [..]@tag"
It doesn't set the vcs.* settings when compiled with go install [..]@tag
or [..]@hash; these are only set when building from a source directory.

Always adding the date would be nice, but this isn't available. There's
an issue on the Go issue tracker (which I can't find right now) that
said that stripping the vcs info is intentional

So use just the tag name or "hash/date" if there is no tag name. We need
to parse the "v2.0.1-0.20251226200708-fefa60ba4a0a" version string when
using [..]@hash.